Move tests into lib
